Add an experiment with multiple services (blue demo)
Problem: I just added a webcache to enhance the performance of the
ratings service.  I want to test it but I don't yet want to deploy it to
production, and I don't want to replicate the existing ratings service
in my test environment.  Besides, whenever I run ratings locally, it's
pretty fast - seems to only get slow in production.

Solution: Add a new experiment template for this - I'll define an
experimental deployment of the services I'm interested in and then use
AspenMesh & Istio to route traffic to my deployments.